Var expansion, task env & obj orig name
For every task object, save the object's original name
as used in the task.

Object entity keeps only the real name of the file,
while different tasks can refer to it by different names.
Store the file name from the task definition in task-object
relation and use it for the expansion of the commands.